piano.c: maybe fix sound

+++ src/piano.c
@@ -1,5 +1,5 @@
 #include <fcntl.h> // For open()
-#include <linux/kd.h> // For KIOCSOUND
+//#include <linux/kd.h> // For KIOCSOUND
 #include <math.h> // For pow()
 #include <stdlib.h> // For atoi()
 #include <sys/ioctl.h> // For ioctl()
@@ -20,7 +20,7 @@ int main(int argc, char *argv[]) {
 	if(argc >= 4) {
 		delay.tv_nsec = atoi(argv[3]) * 1000000;
 	}
-	ioctl(console, KIOCSOUND, (int)(1193180 / (pow(2, (key - 49) / 12) * a4)));
+	ioctl(console, KIOCSOUND, (int)(1193180 / (pow(2, (float)(key - 49) / 12) * a4)));
 	nanosleep(&duration, NULL);
 	ioctl(console, KIOCSOUND, 0);
 	nanosleep(&delay, NULL);
